repo: add missing dependencies and fix churn_library.py format
The matplotlib and seaborn dependencies were missing from the requirements.txt file so have been added here.

Also, the churn_library.py file that was downloaded from Udacity had formating issues when using tab, so the file was deleted here and an updated version will be added in the next commit.