feat: add import_data, create_churn_feature and perform_eda methods
Here the initial code developed for importing the data, create the Churn feature and performing the EDA were added. Note that tehse were included in a class called CustomerChurn.

The tests for each of these methods were also included, along with a confi.yaml file for easily modifying the settings.